requirements.txt
aiohttp
black
coverage
flake8
mock
mysql-connector-python
pyflakes
pylint
pytest
pytest-asyncio
pytest-cov
python-dotenv
requests
requests-toolbelt
ruff
mypy
pre-commit